Added ability to change enum to behave as flags
This required the property to be re-created, since this option affects
the property type. This already happened for class members anyway, but
not for top-level enum properties.

Also fixed an issue with restoring member values after types are changed

When changing this option, existing values do get messed up, because the
integer index gets interpreted as flags, and vice versa. The value is
even entirely lost when the combined flags result in an out-of-range
index. It could be worth looking into improving this behavior.